+++ app/src/main/java/com/marsool/firetool/SharedPrefManager.java
@@ -4,6 +4,8 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 
+import com.marsool.firetool.activities.MainActivity;
+
 /**
  * Created by Belal on 9/5/2017.
  */
