+++ app/src/main/java/com/marsool/firetool/countrycodes/CountryHandler.java
@@ -1,0 +1,5 @@
+package com.marsool.firetool.countrycodes;
+
+public interface CountryHandler {
+    void handle(CountryCode code);
+}
