+++ app/src/main/res/drawable/cadre_bottom.xml
@@ -2,7 +2,7 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
 
     <item
-        android:top="1dp"
+        android:top="-1dp"
         android:bottom="@dimen/border_width"
         android:left="@dimen/border_width"
         android:right="@dimen/border_width">
