+++ app/src/main/res/drawable/cadre_middle.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
+
     <item
-        android:bottom="@dimen/border_width"
+        android:top="-1dp"
+        android:bottom="@dimen/no_border_width"
         android:left="@dimen/border_width"
-        android:right="@dimen/no_border_width"
-        android:top="@dimen/border_width">
+        android:right="@dimen/border_width">
         <shape android:shape="rectangle">
             <solid android:color="@color/input_back" />
-            <corners android:radius="@dimen/corner_radius"/>
             <stroke
                 android:width="@dimen/border_width"
                 android:color="@color/accent" />
