--- app/src/main/res/drawable/cadre_right.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
-    <item
-        android:bottom="@dimen/border_width"
-        android:left="@dimen/no_border_width"
-        android:right="@dimen/border_width"
-        android:top="@dimen/border_width">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/input_back" />
-            <corners android:radius="@dimen/corner_radius"/>
-            <stroke
-                android:width="@dimen/border_width"
-                android:color="@color/accent" />
-        </shape>
-    </item>
-
-</layer-list>
