+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/background_clear"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/content"
@@ -66,8 +66,8 @@
 
             <ImageView
                 android:id="@+id/logo"
-                android:layout_width="150dp"
-                android:layout_height="200dp"
+                android:layout_width="130dp"
+                android:layout_height="180dp"
                 android:layout_marginEnd="24dp"
                 android:background="@drawable/ic_main"
                 android:contentDescription="@string/logo_description"
@@ -82,52 +82,111 @@
             android:id="@+id/constraintLayout"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="32dp"
             android:layoutDirection="inherit"
             android:padding="20dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent">
 
-            <EditText
-                android:id="@+id/editTextUsername"
-                android:layout_width="0dp"
-                android:layout_height="@dimen/input_height"
-                android:layout_weight="1"
-                android:autofillHints=""
+            <LinearLayout
+                android:id="@+id/phone_parent"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
                 android:background="@drawable/cadre_top"
-                android:ems="10"
-                android:hint="@string/phone_field"
-                android:inputType="phone|textNoSuggestions"
-                android:layoutDirection="inherit"
-                android:padding="15dp"
-                android:paddingStart="25dp"
-                android:paddingEnd="25dp"
-                android:textAlignment="viewStart"
+                android:layoutDirection="ltr"
+                android:orientation="horizontal"
                 app:layout_constraintBottom_toTopOf="@+id/editTextPassword"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintHorizontal_bias="0.0"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintVertical_bias="1.0" />
+                app:layout_constraintStart_toStartOf="parent">
+
+                <LinearLayout
+                    android:id="@+id/country"
+                    android:layout_width="wrap_content"
+                    android:layout_height="@dimen/input_height"
+                    android:layout_marginStart="20dp"
+                    android:gravity="center"
+                    android:layoutDirection="inherit"
+                    android:orientation="horizontal">
+
+                    <ImageView
+                        android:id="@+id/country_image"
+                        android:layout_width="25dp"
+                        android:layout_height="25dp"
+                        android:layout_weight="1"
+                        android:scaleType="fitCenter" />
+
+                    <TextView
+                        android:id="@+id/country_code"
+                        android:layout_width="55dp"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="10dp"
+                        android:layout_weight="1"
+                        android:layoutDirection="ltr"
+                        android:textAlignment="textStart"
+                        android:textColor="#000000"
+                        android:textSize="15sp" />
+
+                </LinearLayout>
+
+                <EditText
+                    android:id="@+id/editTextUsername"
+                    android:layout_width="0dp"
+                    android:layout_height="@dimen/input_height"
+                    android:layout_weight="1"
+                    android:autofillHints=""
+                    android:background="@android:color/transparent"
+                    android:ems="10"
+                    android:hint="@string/phone_field"
+                    android:inputType="phone|textNoSuggestions"
+                    android:layoutDirection="inherit"
+                    android:paddingLeft="15dp"
+                    android:paddingRight="25dp"
+                    android:textAlignment="viewStart" />
+            </LinearLayout>
 
             <EditText
                 android:id="@+id/editTextPassword"
                 android:layout_width="0dp"
                 android:layout_height="@dimen/input_height"
-                android:layout_marginBottom="16dp"
                 android:autofillHints=""
-                android:background="@drawable/cadre_bottom"
+                android:background="@drawable/cadre_middle"
                 android:ems="10"
                 android:hint="@string/password_field"
                 android:inputType="textPassword"
+                android:layoutDirection="inherit"
                 android:padding="15dp"
                 android:paddingStart="25dp"
                 android:paddingEnd="25dp"
                 android:textAlignment="viewStart"
-                app:layout_constraintBottom_toTopOf="@+id/message"
+                app:layout_constraintBottom_toTopOf="@id/oneortwo"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent" />
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:id="@+id/oneortwo"
+                android:layout_width="0dp"
+                android:layout_height="@dimen/input_height"
+                android:layout_marginBottom="16dp"
+                android:background="@drawable/cadre_bottom"
+                android:paddingStart="25dp"
+                android:paddingEnd="25dp"
+                app:layout_constraintBottom_toTopOf="@+id/message"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintStart_toStartOf="parent">
+
+                <Spinner
+                    android:id="@+id/number"
+                    style="@style/Widget.AppCompat.Spinner.Underlined"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:dropDownWidth="match_parent"
+                    android:entries="@array/one_or_two"
+                    android:spinnerMode="dropdown"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
 
             <TextView
                 android:id="@+id/message"
@@ -175,7 +234,8 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/connect"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:textAlignment="viewStart">
 
         <ProgressBar
             android:id="@+id/connect_progress"
