+++ app/src/main/res/layout/activity_verify_phone.xml
@@ -8,7 +8,7 @@
     android:layoutDirection="ltr"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
     <RelativeLayout
         android:id="@+id/imageView"
