+++ app/src/main/res/values-en/strings.xml
@@ -13,4 +13,7 @@
     <string name="enter_phone">Please enter your phone number</string>
     <string name="enter_password">Please enter your password</string>
     <string name="about">About</string>
+    <string name="select_country">Select country</string>
+    <string name="exit_confirm">Are you sure you want to quit?</string>
+    <string name="exit">Exit</string>
 </resources>
