+++ app/src/main/res/values/colors.xml
@@ -10,5 +10,5 @@
     <color name="accent">@color/main</color>
     <color name="overlay_light">#ddffffff</color>
     <color name="app_back">#ffffffff</color>
-    <color name="input_back">#88dddddd</color>
+    <color name="input_back">#99eeeeee</color>
 </resources>
